nearbyhint: Fix for ffast-math
The fix on 45cad81a305b09c4b1c9db55c23fcbb0f4e01ee4 wasn't working on
Clang.

On ffast-math the compiler is free to assume that "x + v -v = x".
45cad81a305 was workarounding this fact by storing "x + v" on a volatile
variable.

For Clang this wasn't enough to stop optimizing, as it correctly
detected that the variable is local-scope, so no one can take a
reference to it.

This commit reworks the fix by defining a function to do the operation
and disabling optimizations on that function for all supported
compilers (and those using the same frontend).

For non-supported compilers an #error is emitted, as the workaround
wasn't safe enough. It could even break between compiler versions. This
avoids potentially weird behaviour on the future.