Prevent building invalid batch due to implicit conversion
It is tempting to build a batch from a memory address, but that's not supported,
load_aligned or load_unaligned provide that feature instead.

As C++ conversion rule would still accept that constructor as T* is convertible
to bool which is convertible to batch_bool which is convertible to batch,
prevent construction of batch_bool from pointer.

Note that just using batch(T const*) = delete is not great, as it prevents
initializing a batch<short>(0) : in that case 0 can be converted to short.

Fix #602