Get rid of simd_return_type
This type is used to turn char into either uint8 or int8 depending on its sign.
Unfortunately it is not used uniformly across the code base, so nothing prevents
a user from instantiating a xsimd::batch<char>, and then calling
xsimd::load_aligned from a char pointer, which may resulting in to different
batch types (!).

Icing on the cake: It was misused (not forwarding the Arch parameter) in many case