Fix dispatching mechanism
Traverse required arch in the order provided by the user instead of
trying to guess the best one.

It is actually impossible to define the notion of a best architectures
as Intel instruction sets have a tree structure and not a linear
structure : there are multiple leaves and none of them can be considered
the "best".